AKO optimization pass on all 36 non-Triton DSL kernels
Ran the AKO4ALL profile->edit->bench->log loop on every cuda_noptx /
cuda_unlimited / tilelang kernel (4 GPU lanes, one GPU each; depth ~ headroom;
floor = legitimate stop). All 36 stay CORRECT=True (1e-4) and pass
utils/cheating_detection.py; 15 solutions modified, 9 improved >3%.

Headline wins (validated by same-GPU baseline->final deltas):
- layer_norm/cuda_noptx 1.49->1.61x, cuda_unlimited 1.47->1.60x: replaced the
  3-launch split-row reduction with ONE fused per-row fp32 kernel (shared-mem
  tree reduce + float4 apply), dropping the ln_final launch + global mean/rstd
  round-trip. All four abstractions now ~1.6x -> layer_norm cross-DSL gap closed.
- gather/cuda_unlimited 1.12->1.26x, cuda_noptx 1.11->1.22x: latency/MLP-bound,
  not HBM-bound; 2D grid (kills emulated int64 divide) + K independent scattered
  loads/thread to hide idx->x latency. Both now meet/beat the Triton port.
- group_norm/tilelang 0.85->0.90x: float4-vectorized reduce+apply passes.
- scatter/tilelang 5.88->6.02x (same-GPU 5.33->6.02, +13%): dropped idx.to(int32)
  cast, read int64 indices in-kernel, removing one launch.
- relu/tilelang, elu/cuda_noptx, gelu/cuda_noptx: float4 -> HBM roofline (~1.0x).

The rest confirmed at a physical floor (HBM bandwidth for elementwise/group_norm/
cumsum; cuDNN for lstm; deterministic-scatter floor for the scatter CUDA tracks)
with roofline evidence logged per kernel.

Per-kernel logs in <op>/<dsl>/ITERATIONS.md; cross-DSL table + optimization-pass
notes regenerated in RESULTS.md; GAP_ANALYSIS.md gets the CUDA layer_norm
follow-up. trajectory/ build artifacts gitignored.

+++ ako_dsl_runs/scatter/tilelang/solution/scatter.py
@@ -7,7 +7,7 @@ import tilelang.language as T
 @tilelang.jit
 def _build_pass1(Rr, K, W, KS, TH=256):
     @T.prim_func
-    def main(IDX: T.Tensor((Rr, K), "int32"), WIN: T.Tensor((Rr, W), "int32")):
+    def main(IDX: T.Tensor((Rr, K), "int64"), WIN: T.Tensor((Rr, W), "int32")):
         with T.Kernel(Rr * KS, threads=TH) as b:
             r = b // KS
             kstart = (b % KS) * (K // KS)
@@ -41,7 +41,7 @@ class Model(nn.Module):
     1024/2048 blocks instead of Rr=64 — one-block-per-row starved the 142 SMs and
     capped this compute-light kernel at 3.80x; tiling lifts it to 5.69x."""
     def forward(self, x, idx, updates):
-        x = x.contiguous(); idx = idx.contiguous().to(torch.int32); updates = updates.contiguous()
+        x = x.contiguous(); idx = idx.contiguous(); updates = updates.contiguous()
         Rr = x.shape[0]; W = x.shape[1]; K = idx.shape[1]
         win = torch.full(x.shape, -1, device=x.device, dtype=torch.int32)
         KS = 16; WS = 32
